tests: rename test class
